feat: make `BuildError` newtype
