+++ crates/rolldown_error/src/error_kind/external_entry.rs
