+++ crates/rolldown_error/src/error_kind/unresolved_entry.rs
